+++ src/Demo/Logitar.Demo.Ui/DatabaseProvider.cs
@@ -4,5 +4,6 @@ internal enum DatabaseProvider
 {
   EntityFrameworkCorePostgreSQL,
   EntityFrameworkCoreSqlServer,
-  InMemory
+  InMemory,
+  MongoDB
 }