+++ src/Demo/Logitar.Demo.Ui/Models/Todo/UpdateTodoPayload.cs
@@ -1,0 +1,7 @@
+﻿namespace Logitar.Demo.Ui.Models.Todo;
+
+public record UpdateTodoPayload
+{
+  public string? Text { get; set; }
+  public bool? IsDone { get; set; }
+}