+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implemented query JOINs.
 - Added AggregateRoot metadata.
 - Created a struct for actor identifiers.
+- Implemented an EventSourcing MongoDB store.
+- Added a demo TodoController with CRUD endpoints.
 
 ### Changed